moved EC2InstanceBootSpecific to aws-library
